feat: Enhance text rendering and rectangle management
- Added a `set` method to the `Rectangle` class for easier manipulation of rectangle properties.
- Integrated `CanvasTextGenerator` into the text rendering pipeline to improve text measurement and rendering efficiency.
- Updated `CanvasTextSystem` to utilize the new `CanvasTextGenerator` for texture creation, enhancing performance and reducing code complexity.
- Adjusted text bounds updating logic to account for trimming, ensuring accurate positioning of text on screen.
- Improved the `getCanvasBoundingBox` function to allow for custom dimensions and resolutions, optimizing performance for large canvases.
